move test to dedicated file
